+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/SemverPlugin.kt
@@ -5,6 +5,7 @@ import com.javiersc.semver.gradle.plugin.services.GitBuildService
 import com.javiersc.semver.gradle.plugin.tasks.CreateSemverTagTask
 import com.javiersc.semver.gradle.plugin.tasks.PrintSemverTask
 import com.javiersc.semver.gradle.plugin.tasks.PushSemverTagTask
+import com.javiersc.semver.gradle.plugin.tasks.WriteSemverTask
 import com.javiersc.semver.gradle.plugin.valuesources.VersionValueSource
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -31,6 +32,7 @@ public class SemverPlugin : Plugin<Project> {
         CreateSemverTagTask.register(this, gitTagBuildService)
         PushSemverTagTask.register(this, gitTagBuildService)
         PrintSemverTask.register(this)
+        WriteSemverTask.register(this)
     }
 
     private fun Project.configureLazyVersion(gitTagBuildService: Provider<GitBuildService>) {