+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/tasks/PrintSemverTask.kt
@@ -1,5 +1,6 @@
 package com.javiersc.semver.gradle.plugin.tasks
 
+import com.javiersc.gradle.project.extensions.isRootProject
 import com.javiersc.gradle.tasks.extensions.maybeRegisterLazily
 import com.javiersc.gradle.tasks.extensions.namedLazily
 import com.javiersc.semver.gradle.plugin.internal.semverMessage
@@ -8,43 +9,95 @@ import javax.inject.Inject
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
 import org.gradle.api.Task
-import org.gradle.api.file.ProjectLayout
-import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
-import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.TaskProvider
+import org.gradle.api.tasks.UntrackedTask
+import org.gradle.api.tasks.options.Option
 import org.gradle.api.tasks.testing.Test
 import org.gradle.jvm.tasks.Jar
+import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
 import org.gradle.kotlin.dsl.withType
 import org.gradle.language.base.plugins.LifecycleBasePlugin.ASSEMBLE_TASK_NAME
 import org.gradle.language.base.plugins.LifecycleBasePlugin.BUILD_TASK_NAME
 import org.gradle.language.base.plugins.LifecycleBasePlugin.CHECK_TASK_NAME
+import org.gradle.process.ExecOperations
 
-@CacheableTask
+@UntrackedTask(because = "It must always print the version")
 public abstract class PrintSemverTask
 @Inject
 constructor(
+    private val isRootProject: Boolean,
     private val projectName: String,
     objects: ObjectFactory,
-    layout: ProjectLayout,
+    private val execOperations: ExecOperations,
 ) : DefaultTask() {
 
     init {
         group = "semver"
     }
 
+    @Input
+    @Option(
+        option = "githubOnlyRoot",
+        description = "Set any GitHub option to work only on the root project `printSemver` task"
+    )
+    public val githubOnlyRoot: Property<Boolean> = objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubEnvTag",
+        description =
+            "Set the version as `semver-tag-subproject-name` environment variable of the GitHub Actions"
+    )
+    public val githubEnvTag: Property<Boolean> = objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubEnvVersion",
+        description =
+            "Set the version as `semver-version-subproject-name` environment variable of the GitHub Actions"
+    )
+    public val githubEnvVersion: Property<Boolean> = objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubEnv",
+        description =
+            "Set the version as `semver-subproject-name` environment variable of the GitHub Actions"
+    )
+    public val githubEnv: Property<Boolean> = objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubOutputTag",
+        description = "Set the version as `semver-tag-subproject-name` output of the GitHub Actions"
+    )
+    public val githubOutputTag: Property<Boolean> = objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubOutputVersion",
+        description =
+            "Set the version as `semver-version-subproject-name` output of the GitHub Actions"
+    )
+    public val githubOutputVersion: Property<Boolean> =
+        objects.property<Boolean>().convention(false)
+
+    @Input
+    @Option(
+        option = "githubOutput",
+        description =
+            "Set the version as `semver-subproject-name` output of the GitHub Actions `semver` step ID"
+    )
+    public val githubOutput: Property<Boolean> = objects.property<Boolean>().convention(false)
+
     @get:Input public abstract val tagPrefix: Property<String>
 
     @get:Input public abstract val version: Property<String>
-
-    @OutputFile
-    public val semverFile: RegularFileProperty =
-        objects.fileProperty().convention(layout.buildDirectory.file("semver/version.txt"))
 
     @TaskAction
     public fun run() {
@@ -57,17 +110,36 @@ constructor(
 
         semverMessage("semver for $name: $semverWithPrefix")
 
-        semverFile.get().asFile.apply {
-            parentFile.mkdirs()
-            createNewFile()
-            writeText(
-                """
-                    |$semver
-                    |$semverWithPrefix
-                    |
-                """.trimMargin()
-            )
+        val onlyRoot: Boolean = githubOnlyRoot.orNull ?: false
+        val allProjects: Boolean = !onlyRoot
+        when {
+            onlyRoot && isRootProject -> configureGitHub(prefix, semver, semverWithPrefix)
+            allProjects -> configureGitHub(prefix, semver, semverWithPrefix)
         }
+    }
+
+    private fun configureGitHub(tagPrefix: String, semver: String, semverWithPrefix: String) {
+        val tagName = if (isRootProject) "semver-tag" else "semver-tag-$projectName"
+        if (githubOutputTag.orNull == true) executeGitHubOutput(tagName, tagPrefix)
+        if (githubEnvTag.orNull == true) executeGitHubEcho(tagName, tagPrefix)
+
+        val versionName = if (isRootProject) "semver-version" else "semver-version-$projectName"
+        if (githubOutputVersion.orNull == true) executeGitHubOutput(versionName, semver)
+        if (githubEnvVersion.orNull == true) executeGitHubEcho(versionName, semver)
+
+        val name = if (isRootProject) "semver" else "semver-$projectName"
+        if (githubOutput.orNull == true) executeGitHubOutput(name, semverWithPrefix)
+        if (githubEnv.orNull == true) executeGitHubEcho(name, semverWithPrefix)
+    }
+
+    private fun executeGitHubOutput(key: String, value: String) {
+        semverMessage("\nSetting $value as `$key` output:")
+        println("::set-output name=$key::$value")
+    }
+
+    private fun executeGitHubEcho(key: String, value: String) {
+        semverMessage("\nSetting $value as `$key` environment variable:")
+        execOperations.exec { commandLine("echo", "$key=$value", ">> \$GITHUB_ENV") }
     }
 
     internal companion object {
@@ -75,9 +147,10 @@ constructor(
 
         fun register(project: Project): TaskProvider<PrintSemverTask> {
             val printSemverTask: TaskProvider<PrintSemverTask> =
-                project.tasks.register(taskName, project.name)
+                project.tasks.register(taskName, project.isRootProject, project.name)
 
             printSemverTask.configure {
+                dependsOn(WriteSemverTask.taskName)
                 tagPrefix.set(project.semverExtension.tagPrefix)
                 version.set(project.version.toString())
             }